% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.branch.lengths.R,
%   R/phylogenetic.distabnce.R
\name{add.branch.lengths}
\alias{add.branch.lengths}
\title{Add branch lengths to a phylogenetic tree}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]

Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
add.branch.lengths(tree, rounding = 3, cex = 0.5, ...)

add.branch.lengths(tree, rounding = 3, cex = 0.5, ...)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{rounding}{A number of significant digits for rounding of the branch lengths. Defaults to 3.}

\item{cex}{A number for the graphic scaling factor of the text for branch lengths. Defaults to 0.5.}
}
\description{
Add branch lengths to a phylogenetic tree

Add branch lengths to a phylogenetic tree
}
\examples{
plot(anole.tree)
add.branch.lengths(anole.tree)
plot(anole.tree)
add.branch.lengths(anole.tree)
}
