% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.sequences.R
\name{align.sequences}
\alias{align.sequences}
\title{Align DNA or amino acid sequences}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.sequences(x, method = "ClustalOmega", verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of biological sequence data. Must be one of the following:
A character string containing the FASTA format sequences as downloaded from NCBI.
A named vector of sequences.
A data frame with columns for the sequence names and sequence text, as produced by \code{phylotools::read.fasta}.
A \code{Biostrings::DNAStringSet} or \code{AAStringSet} object.}

\item{method}{The alignment algorithm. Options include \code{"ClustalOmega"} (default), \code{"ClustalW"} and \code{"Muscle"}.}

\item{verbose}{A logical argument specifying whether to display confirmation. Defaults to \code{TRUE}.}
}
\value{
Returns an \code{MsaDNAMultipleAlignment} or \code{MsaAAMultipleAlignment} object.
}
\description{
Align DNA or amino acid sequences
}
\details{
This function is intended as a user-friendly wrapper for \code{msa::msa} and accepts other arguments to that function.
}
\examples{
anole.alignment <- align.sequences(anole.ND2)
print(anole.alignment, show="complete")

}
