% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anole.natural.history}
\alias{anole.natural.history}
\title{A dataset of anole species, ecomorphs and their islands of origin}
\format{
An object of class \code{data.frame} with 17 rows and 3 columns.
}
\source{
Nicholson, K. E., Harmon, L. J. and Losos, J. B. 2007. Evolution of \emph{Anolis} Lizard Dewlap Diversity. \emph{Evolution} 2(3): e274
    \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1803026/}
}
\usage{
anole.natural.history
}
\description{
Data for 16 species of \emph{Anolis} lizard and the related lizard \emph{Leiocephalus barahonensis},
    listing their endemic island and ecomorph.
}
\examples{
unique(anole.natural.history$island)
unlist(lapply(unique(anole.natural.history$island), function(x) {
  sum(anole.natural.history$island==x)
}))
unique(anole.natural.history$ecomorph)
unlist(lapply(unique(anole.natural.history$ecomorph), function(x) {
  sum(anole.natural.history$ecomorph==x)
}))

}
\keyword{datasets}
