% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anole.ND2}
\alias{anole.ND2}
\title{Anole DNA sequences}
\format{
An object of class \code{DNAStringSet} of length 17.
}
\usage{
anole.ND2
}
\description{
Unaligned sequences from mitochondrial DNA of 17 species of anoles, including
    the complete CDS of NADH dehydrogenase subunit 2 (ND2), the complete tRNA-Trp sequence,
    partial sequence from tRNA-Ala.
}
\examples{
anole.ND2
names(anole.ND2)

# Make the names more minimalist
names(anole.ND2) <- sub("_ND2-tRNAtrp-tRNAala_","__",names(anole.ND2))

# Par the names down to just the species
names(anole.ND2) <- stringr::str_split_fixed(names(anole.ND2),"__",2)[,1]

# Alignment
require(msa)
seqs.msa <- msa(anole.ND2, method = "ClustalOmega", order = "aligned")
print(seqs.msa, show="complete")
}
\keyword{datasets}
