% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anole.tree}
\alias{anole.tree}
\title{Phylogenetic tree of anole species}
\format{
An object of class \code{phylo} of length 4.
}
\usage{
anole.tree
}
\description{
A phylogram produced by neighboring joining of anole ND2 mitochondrial sequences (the \code{anole.ND2} dataset).
}
\examples{
anole.tree
plot(anole.tree)
}
\keyword{datasets}
