% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.sequences.R
\name{combine.sequences}
\alias{combine.sequences}
\title{Combine two sequence data objects}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
combine.sequences(x1, x2, verbose = TRUE, ...)
}
\arguments{
\item{x1, x2}{Objects of biological sequence data.}

\item{verbose}{A logical argument specifying whether to display the summary output. Defaults to \code{TRUE}.}
}
\value{
Returns one sequence data object in the same format as the two that are provided. Sequences in `x1` will appear before those in `x2`.
}
\description{
Combine two sequence data objects of the same type. The objects must be one of the following:
    1. Character strings containing FASTA format sequences as downloaded from NCBI.
    2. Character vectors of sequences.
    3. Data frames with columns for the sequence names and sequence text, e.g. as produced by \code{phylotools::read.fasta}.
    4. \code{Biostrings::DNAStringSet} or 5. \code{AAStringSet} objects.
}
\examples{
accession.ids <- c("EF591774.1", "AF055971.2", "AF055967.2", "AF055927.2", "AY296172.1", "AF294303.1", "AF294297.1", "AF055976.2", "AF055945.2", "AF294317.1", "AY296196.1", "AF055966.2", "AF055939.2", "AF055940.2", "AY296195.1", "AY263052.1", "AY296163.1")
anole.nd2 <- combine.sequences(accession.ids)
}
