% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dewlaps}
\alias{dewlaps}
\title{A dataset on the dewlaps of anole lizards}
\format{
An object of class \code{data.frame} with 47 rows and 6 columns.
}
\source{
Nicholson, K. E., Harmon, L. J. and Losos, J. B. 2007. Evolution of \emph{Anolis} Lizard Dewlap Diversity. \emph{Evolution} 2(3): e274
    \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1803026/}
}
\usage{
dewlaps
}
\description{
Data for 47 species of \emph{Anolis} lizards. Dewlaps are skin flaps under the chin, which are
    displayed by males to communicate mating and territorial interests. This table comes from Nicholson et al. 2007 (Table 4)
    and lists species, community (island), ecomorph, the relative size of the dewlap, dewlap pattern and color.
    Relative dewlap size are the residuals of actual dewlap size regressed against snout-vent length.
}
\examples{
boxplot(dewlaps$dewlap.relative.size ~ dewlaps$community)
}
\keyword{datasets}
