% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.tree.R
\name{draw.tree}
\alias{draw.tree}
\title{Plot a phylogenetic tree}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
draw.tree(
  tree,
  font = 2,
  tip.color = "black",
  edge.width = 3,
  show.branch.lengths = TRUE,
  branch.length.cex = 0.7,
  branch.length.text.color = "gray50",
  show.scale.bar = TRUE,
  margins = c(0.5, 0.5, 0.5, 0.5),
  ...
)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{font}{An integer specifying the type of font for the labels: 1 (plain text), 2 (bold), 3 (italic, the default), or 4 (bold italic).}

\item{tip.color}{A color name or vector for the tip label text.}

\item{edge.width}{An integer specifying the width of branches in the tree.}

\item{show.branch.lengths}{Logical (TRUE or FALSE) specifying whether to show branch lengths on the tree. (Default is TRUE.)}

\item{branch.length.cex}{An integer specifying the character expansion factor for branch length text. (Default is 0.7)}

\item{branch.length.text.color}{A color name for the branch length text.}

\item{show.scale.bar}{Logical (TRUE or FALSE) specifying whether to show a scale bar. (Default is TRUE.)}

\item{margins}{A vector of four integers defining the margins of the plot.}
}
\description{
This function plots a phylogenetic tree (of class \code{phylo}).
The function is a wrapper for \code{ape::plot.phylo} with a few stylistic changes to the default,
including thicker branches, slightly offset tip labels, different default font style (bold).
It also includes the display of branch lengths by default using \code{ape::edgelabels},
with stylistic changes to make it a bit cleaner.
}
\examples{
draw.tree(anole.tree)

}
