% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.sequence.R
\name{fetch.sequence}
\alias{fetch.sequence}
\title{Download sequence data from NCBI Entrez database based on sequence IDs}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
fetch.sequence(
  ids,
  database = "nuccore",
  format = "DNAStringSet",
  simplify.names = TRUE,
  sequence.name = "",
  separator = "_",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ids}{A character vector of valid NCBI UIDs or accession numbers.}

\item{database}{A character name of an NCBI database. Defaults to \code{"nuccore"}, or the non-redundant nucleotide database (NR).
For a complete list of options run \code{rentrez::entrez_db_searchable()}}

\item{format}{A character name for the output format.
\code{"raw"}: A character string containing the FASTA format sequences as downloaded from NCBI.
\code{"character"}: A named vector of sequences.
\code{"data.frame"} or \code{"table"}: A data frame with columns for the sequence names and sequence text, as produced by \code{phytools::read.fasta}.
\code{"DNAStringSet"}: A \code{Biostrings::DNAStringSet} object. (Default)
\code{"AAStringSet"}: A \code{Biostrings::AAStringSet} object.}

\item{simplify.names}{A logical argument specifying whether to simplify the sequence names to species - sequence name - accession number. Defaults to \code{TRUE}.
The species names and accession numbers are pulled from NCBI. The sequence name must be provided by the user.}

\item{sequence.name}{An optional sequence name.}

\item{separator}{A character to use to separate the components of simplified sequence names. Defaults to the underscore.}

\item{verbose}{A logical argument specifying whether to display the summary output. Defaults to \code{TRUE}.}
}
\value{
Returns sequence data in the format specified by the \code{format} argument.
}
\description{
Download sequence data from NCBI Entrez database based on sequence IDs
}
\details{
This function is intended as a user-friendly wrapper for \code{rentrez::entrez_fetch} and accepts other arguments to that function.
}
\examples{
accession.ids <- c("EF591774.1", "AF055971.2", "AF055967.2", "AF055927.2", "AY296172.1", "AF294303.1", "AF294297.1", "AF055976.2", "AF055945.2", "AF294317.1", "AY296196.1", "AF055966.2", "AF055939.2", "AF055940.2", "AY296195.1", "AY263052.1", "AY296163.1")
anole.nd2 <- fetch.sequence(accession.ids)
}
