% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finches}
\alias{finches}
\title{Data on beak depth in Darwin's finches}
\format{
An object of class \code{data.frame} with 82 rows and 3 columns.
}
\source{
Boag, P. T. and Grant, P. R. 1981. Intense Natural Selection in a Population of Darwin's Finches (Geospizinae) in the Galápagos.
    \emph{Science} 214(4516): 82-85.
    \url{https://doi.org/10.1126/science.214.4516.82}
}
\usage{
finches
}
\description{
Data on the distributions of beak depth in Darwin's medium ground finch (*Geospiza fortis*) from Daphne Major in
    the Galápagos islands for the years 1976 and 1978, flanking a drought year in 1977.
    These data are inferred from the histogram appearing in Freeman's *Biological Science*, 7th edition,
    Figure 22.12, which are based on Boag & Grant (1981).
    They are intended as an illustration of the effects of intense natural selection on a quantitative trait.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)

# sample size by year
finches \%>\%
  group_by(year) \%>\%
  summarize(sample.size = sum(count))

# means by year
mean.beaks <- finches \%>\%
  filter(count!=0) \%>\%
  mutate(sum.of.depths = beak_depth_mm*count) \%>\%
  group_by(year) \%>\%
  summarize(mean = sum(sum.of.depths)/sum(count))

# SD by year
beak.sd <- finches \%>\%
  mutate(mean = c(rep(mean.beaks$mean[1],dim(finches)[1]/2),rep(mean.beaks$mean[2],dim(finches)[1]/2))) \%>\%
  mutate(squared.deviation = (beak_depth_mm - mean)^2) \%>\%
  mutate(SSD.by.category = squared.deviation*count) \%>\%
  filter(count!=0) \%>\%
  group_by(year) \%>\%
  summarize(
    sd = sqrt(sum(SSD.by.category)/(sum(count)-1)),
    mean = unique(mean)
  ) \%>\%
  mutate(
    meanPlus1SD = mean + sd,
    meanPlus2SD = mean + 2*sd,
    meanMinus1SD = mean - sd,
    meanMinus2SD = mean - 2*sd
  )

finches \%>\%
  ggplot(aes(x=beak_depth_mm, y=count)) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  ) +
  facet_grid(year~.) + # scales = "free_y"
  geom_col(width = 0.2) +
  geom_hline(yintercept = 0, size = 0.5, color = "gray50") +
  geom_vline(data = mean.beaks, aes(xintercept = mean, group = year), color = "darkred") +
  geom_vline(data = beak.sd, aes(xintercept = meanPlus1SD, group = year), color = "gray20", linetype = 3) +
  geom_vline(data = beak.sd, aes(xintercept = meanPlus2SD, group = year), color = "gray20", linetype = 3) +
  geom_vline(data = beak.sd, aes(xintercept = meanMinus1SD, group = year), color = "gray20", linetype = 3) +
  geom_vline(data = beak.sd, aes(xintercept = meanMinus2SD, group = year), color = "gray20", linetype = 3) +
  labs(x = "beak depth (mm)", y = "number of finches", caption = "Data from Boag & Grant (1981)")

ggsave('finches.png', width = 4, height = 5, scale = 1)

}
\keyword{datasets}
