% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish.COII}
\alias{fish.COII}
\title{Fish cytochrome c oxidase subunit II amino acid sequences}
\format{
An object of class \code{AAStringSet} of length 5.
}
\usage{
fish.COII
}
\description{
Unaligned sequences from mitochondrial cytochrome c oxidase subunit II
    from shark, coelacanth, lungfish, zebrafish and yellow-bellied toad.
}
\examples{
fish.COII
names(fish.COII)

# Alignment
require(msa)
seqs.msa <- msa(fish.COII, type = "protein", method = "ClustalOmega", order = "input")
print(seqs.msa, show="complete")
}
\keyword{datasets}
