% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genome.size}
\alias{genome.size}
\title{Data on genome sizes}
\format{
An object of class \code{data.frame} with 509 rows and 10 columns.
}
\usage{
genome.size
}
\description{
A dataset including genome size (in base pairs), predicted gene number, percent G+C content and
    haploid chromosome number for more than 500 species.
}
\details{
Data are from NCBI where available with many details filled in from the primary literature.
    This dataset is for teaching pursues. Its accuracy and currency cannot be maintained.
}
\examples{
library(ggplot2)
library(ggrepel)
library(ggpubr)

scatterplot.theme <- theme_bw() + theme(
  plot.title = element_text(size=16,face="bold",hjust=0.5),
  axis.text.x = element_text(size=12,face="plain"),
  axis.text.y = element_text(size=12,face="plain"),
  axis.title.x = element_text(size=14,face="plain"),
  axis.title.y = element_text(size=14,face="plain"),
  panel.border = element_rect(fill = NA, color = "black")
)

taxon.colors=c("darkolivegreen3", "darkred", "brown", "#AA336A", "darkgrey", "antiquewhite4", "darkblue", "darkgoldenrod", "black")

genome.fig1 <- ggplot(genome.size) +
  scatterplot.theme + theme(legend.position="none") +
  geom_point(aes(x=log10(genome_size), y=log10(gene_number), colour = taxon, alpha = 0.9), size=3) +
  geom_text_repel(aes(x=log10(genome_size), y=log10(gene_number), label = fig1_labels),
                  size =3.5, max.iter = 50000, max.overlaps = 100) +
  scale_colour_manual(values = taxon.colors) +
  scale_alpha(guide = 'none') +
  xlab(expression(paste(log[10]," genome size (bp)", sep=""))) +
  ylab(expression(paste(log[10]," gene number", sep="")))

genome.fig2 <- ggplot(genome.size) +
  scatterplot.theme +
  geom_point(aes(x=log10(genome_size), y=log10(gene_number), colour = taxon, alpha = 0.9), size=3) +
  geom_text_repel(aes(x=log10(genome_size), y=log10(gene_number), label = fig2_labels),
                  size =3.5, max.iter = 50000, max.overlaps = 1000) +
  scale_colour_manual(values = taxon.colors) +
  scale_alpha(guide = 'none') +
  coord_cartesian(xlim = c(7.5, 11.25), ylim = c(4, 5.25)) +
  xlab(expression(paste(log[10]," genome size (bp)", sep=""))) +
  ylab(expression(paste(log[10]," gene number", sep="")))

genome.figs <- ggarrange(genome.fig1, genome.fig2, ncol=2, nrow=1,
                         widths = c(1,1.5))

genome.figs

ggsave('genome.size.figure.jpg', genome.figs, width=9,height=6, scale=1.25)

}
\keyword{datasets}
