% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.chrs}
\alias{human.chrs}
\title{Gene content of human chromosomes}
\format{
An object of class \code{data.frame} with 25 rows and 13 columns.
}
\usage{
human.chrs
}
\description{
The dataset includes information on the size and gene content of each human chromosomes, including
    NCBI RefSeq identifiers, length in base pairs, the numbers of annotated protein-coding genes, pseudogenes from GRCh37 release 109,
    the average G+C percentage, the numbers of annotated ribosomal RNAs (rRNAs), transfer RNAs (tRNAs), microRNAs,
    small nuclear RNAs (snRNAs), small nucleolar RNAs (snoRNAs), and long non-coding RNAs (lncRNAs) from GRCh37 release 110.
}
\examples{
library(magrittr)
library(dplyr)
library(stringr)
library(tidyr)
library(ggplot2)
library(ggpubr)

chr.length <- human.chrs \%>\%
  mutate(chr.number = str_pad(1:25, width = 2, pad = "0")) \%>\%
  ggplot(aes(x = chr.number, y = length_bp/1e6)) +
  theme_bw() +
  theme( panel.grid.major.x = element_blank() ) +
  geom_bar(stat='identity') +
  scale_x_discrete(name = "chromosome", labels = human.chrs$chr) +
  scale_y_continuous(name = "length (Mbp)")

length.v.genes <- human.chrs \%>\%
  mutate(chr.number = str_pad(1:25, width = 2, pad = "0")) \%>\%
  ggplot(aes(x = length_bp/1e6, y = protein_coding_genes, label = chr)) +
  theme_bw() +
  theme( panel.grid.minor = element_blank() ) +
  geom_smooth(method=lm, fill = "gray85", color = "darkred") +
  geom_text() +
  scale_x_continuous(name = "chromosome length (Mbp)") +
  scale_y_continuous(name = "number of protein-coding genes")

genes.by.chr <- ggarrange(chr.length, length.v.genes, ncol = 2)

ggsave('genes.by.chr.jpg', genes.by.chr, width = 9.5, height = 4.5, scale=1)

rnas.by.chr <- human.chrs \%>\%
  mutate(chr.number = str_pad(1:25, width = 2, pad = "0")) \%>\%
  pivot_longer(cols = grep("RNA",colnames(human.chrs)), names_to = "RNA.type", values_to = "count") \%>\%
  ggplot(aes(x = rev(chr.number), y = count, group = RNA.type)) +
  theme_bw() +
  theme( strip.background = element_blank() ) +
  facet_grid(.~RNA.type, scales = "free") +
  geom_bar(stat = "identity") +
  scale_x_discrete(name = "chromosome", labels = rev(human.chrs$chr)) +
  scale_y_continuous(name = "count") +
  coord_flip()

ggsave('rnas.by.chr.jpg', rnas.by.chr, width = 9.5, height = 4.5, scale=1)

}
\keyword{datasets}
