% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.phylogeny.R
\name{infer.phylogeny}
\alias{infer.phylogeny}
\title{Infer phylogeny for aligned DNA or amino acid sequences}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
infer.phylogeny(
  x,
  method = c("NJ", "UPGMA"),
  model = c("F81", "JC69", "WAG", "JTT", "LG", "Dayhoff", "cpREV", "mtmam", "mtArt",
    "MtZoa", "mtREV24", "VT", "RtREV", "HIVw", "HIVb", "FLU", "Blosum62",
    "Dayhoff_DCMut", "JTT_DCMut"),
  truncate.names = NULL,
  show.tree = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{an \code{MsaDNAMultipleAlignment} or \code{MsaAAMultipleAlignment} object.}

\item{method}{A character indicating the phylogenetic method. Available methods include \code{"NJ"} (default) or \code{"UPGMA"}}

\item{model}{A character indicating the substitution model to be used.
For DNA alignments, available methods include "F81" and "JC69" (default) .
For amino acid alignments: "WAG", "JTT", "LG", "Dayhoff", "cpREV", "mtmam", "mtArt", "MtZoa", "mtREV24", "VT","RtREV", "HIVw", "HIVb", "FLU", "Blosum62" (default), "Dayhoff_DCMut" and "JTT_DCMut".
For more information see \code{?phangorn::dist.ml}.}

\item{truncate.names}{An optional number of characters by which to truncate sequence names.}

\item{show.tree}{A logical argument specifying whether to display the resulting phylogeny. Defaults to \code{TRUE}.}

\item{verbose}{A logical argument specifying whether to display details. Defaults to \code{TRUE}.}
}
\value{
Returns an \code{phylo} tree object.
}
\description{
Infer phylogeny for aligned DNA or amino acid sequences
}
\details{
This function is intended as a user-friendly wrapper for \code{phangorn::dist.ml} and \code{NJ} or \code{upgma} functions.
    Additional arguments will be passed to \code{dist.ml}.
}
\examples{
anole.alignment <- align.sequences(anole.ND2)
anole.tree <- infer.phylogeny(anole.alignment)

}
