% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.clade.R
\name{label.clade}
\alias{label.clade}
\title{Label a clade on the plot of a phylogenetic tree}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
label.clade(
  tree = NULL,
  text,
  clade,
  offset = NULL,
  wing.length = 0,
  cex = 1.2,
  orientation = "vertical",
  font = 2,
  line.width = 7,
  line.color = "darkred",
  text.color = "darkred"
)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{text}{Text that will label the clade.}

\item{clade}{A character vector with two or more taxon names (tip labels) to define the clade.}

\item{offset}{An optional integer value to shift the clade label to the right.}

\item{wing.length}{An optional integer value for the length of top and bottom brackets on the line that highlights the clade.}

\item{cex}{An optional integer value for the character expansion factor to make the text larger or smaller.}

\item{orientation}{The orientation of the text. Can be orientation = "vertical" (the default) or "horizontal".}

\item{font}{An integer specifying the type of font for the labels: 1 (plain text), 2 (bold), 3 (italic, the default), or 4 (bold italic).}

\item{line.width}{An integer specifying the width of branches in the tree.}

\item{line.color}{A color name for the line.}

\item{text.color}{A color name for the label text.}
}
\description{
This function modifies \code{phytools::cladelabels}, written by Liam Revell,
making it easier to identify clades based on tip labels and
to allow users more formatting options.
The names provided to \code{clade} can be any unique component of the taxon names (tip labels).
Only works with rightward facing trees.
}
\examples{
plot.tree(anole.tree)

}
