% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mendel.dihybrid}
\alias{Mendel.dihybrid}
\title{Data from Mendel 1866 on his Dihybrid Cross}
\format{
An object of class \code{data.frame} with 4 rows and 4 columns.
}
\source{
Mendel, G. 1866. Versuche über Plflanzenhybriden. \emp{Verhandlungen des naturforschenden Vereines in Brünn, Bd. IV für das Jahr 1865}, Abhandlungen, 3-47.
    \url{http://www.netspace.org./MendelWeb/}
}
\usage{
Mendel.dihybrid
}
\description{
A dataset reporting the offspring from the dihybrid cross in the garden pea \emph{Pisum sativum} focusing on seed shape and color.
}
\details{
Mendel crossed seed parents with round yellow seeds (the two dominant phenotype for those traits)
    with a pollen parent with wrinkled green seeds (the two recessive phenotypes).
    The F1 offspring were all round and yellow like those of the P1 seed parents.
    Next, 15 plants yielded offspring with all four possible combinations of seed shape and seed color \code{F2_phenotype}.
    The actual numbers are listed in the column \code{F2_count} and appear in the ratio of 9:3:3:1.
    Plants grown from the round yellow seeds were allowed to self-fertilize. These produced several \code{F3_phenotype} classes
    whose numbers are listed in the column \code{F3_count} and reveal a 1:2:2:4 ratio of homozygosity to heterozygosity.
}
\keyword{datasets}
