% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mendel.monohybrid.Exp2}
\alias{Mendel.monohybrid.Exp2}
\title{Data from Mendel 1866 on Experiment 2, the Monohybrid Cross Focusing on Seed Color}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\source{
Mendel, G. 1866. Versuche über Plflanzenhybriden. \emp{Verhandlungen des naturforschenden Vereines in Brünn, Bd. IV für das Jahr 1865}, Abhandlungen, 3-47.
    \url{http://www.netspace.org./MendelWeb/}
}
\usage{
Mendel.monohybrid.Exp2
}
\description{
A dataset reporting the offspring from replicate monohybrid crosses of the garden pea \emph{Pisum sativum} focusing on seed color
}
\details{
The table lists the numbers of F2 plants producing \code{yellow} and \code{green} seeds from the cross of F1 heterozygous parents.
}
\keyword{datasets}
