% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mendel.monohybrid}
\alias{Mendel.monohybrid}
\title{Data from Mendel 1866 on Monohybrid Crosses}
\format{
An object of class \code{data.frame} with 7 rows and 11 columns.
}
\source{
Mendel, G. 1866. Versuche über Plflanzenhybriden. \emp{Verhandlungen des naturforschenden Vereines in Brünn, Bd. IV für das Jahr 1865}, Abhandlungen, 3-47.
    \url{http://www.netspace.org./MendelWeb/}
}
\usage{
Mendel.monohybrid
}
\description{
A dataset reporting the offspring from monohybrid crosses in the garden pea \emph{Pisum sativum} for 7 traits.
}
\details{
The table lists \code{experiment_number} used by Mendel, the \code{focal_trait} for each experiment,
    the \code{number_of_P1_fertilizations} and the \code{number_of_P1_plants}. For the first two
    crosses focusing on seed shape and color the \code{number_of_F1_plants} is listed.
    The \code{dominant_state} and \code{recessive_state} for each trait are listed in separate columns.
    The columns \code{F2_dominant} and \code{F2_recessive} listed the number of F2 offspring displaying the dominant and recessive states for each set of crosses.
    Mendel also allowed F2 plants with the dominant phenotype to self-fertilize.
    The numbers of those plants producing F3 offspring with only the dominant phenotype are given in the columns \code{F3_dominant_only}.
    \code{F3_mixed} lists the number of plants that produced a mix of offspring with the dominant and recessive phenotypes
    ("in the proportion of 3:1," as Mendel put it).
}
\keyword{datasets}
