% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mendel.trihybrid}
\alias{Mendel.trihybrid}
\title{Data from Mendel 1866 on his Trihybrid Cross}
\format{
An object of class \code{data.frame} with 27 rows and 5 columns.
}
\source{
Mendel, G. 1866. Versuche über Plflanzenhybriden. \emp{Verhandlungen des naturforschenden Vereines in Brünn, Bd. IV für das Jahr 1865}, Abhandlungen, 3-47.
    \url{http://www.netspace.org./MendelWeb/}
}
\usage{
Mendel.trihybrid
}
\description{
A dataset reporting the offspring from the trihybrid cross in the garden pea \emph{Pisum sativum} focusing on seed shape, seed color and flower color.
}
\details{
Mendel crossed seed parents with round yellow seeds and violet flowers (the dominant phenotypes for those three traits)
    with a pollen parent with wrinkled green seeds and white flowers (the recessive phenotypes).
    The F1 offspring were all round yellow seeds which grew into plants with violet flowers, like those of the P1 seed parents.
    Next, 24 plants yielded offspring with all possible combinations of seed shape and seed color \code{F2_phenotype}.
    Mendel reported the actual numbers from this cross by already organizing plants into classes (\code{Mendels_class}) based on subsequent self-fertilization to identity heterozygotes.
    This table provides the phenotypes for \code{seed_shape}, \code{seed_color} and \code{flower_color} for each class as well as their \code{count}.
    Combining the counts by phenotype will reveal the expected F2 trihybrid ratio of 27:9:9:9:3:3:3:1.
}
\examples{
library(magrittr)
library(dplyr)

Mendel.trihybrid \%>\%
  mutate(phenotype = paste(seed_shape, seed_color, flower_color, sep="/")) \%>\%
  group_by(phenotype) \%>\%
  summarize(count = sum(count))

}
\keyword{datasets}
