% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.distance.R
\name{phylogenetic.distance}
\alias{phylogenetic.distance}
\title{Find the distance between two tips of a phylogenetic tree}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
phylogenetic.distance(tree, tip1, tip2)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{tip1}{A character exactly matching one tip label in \code{tree}.}

\item{tip2}{A character exactly matching another tip label in \code{tree}.}
}
\description{
Find the distance between two tips of a phylogenetic tree
}
\examples{
plot(anole.tree)
phylogenetic.distance(anole.tree, "A. olssoni","A. cybotes")
phylogenetic.distance(anole.tree, "A. cybotes", "A. occultus")

# To find all pairwise distances
ape::cophenetic.phylo(anole.tree)
}
