% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{Read in a FASTA biological sequence file}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
read.fasta(filename, format = "DNAStringSet", verbose = TRUE, ...)
}
\arguments{
\item{filename}{A character with the path and filename to FASTA file}

\item{format}{A character name for the output format.
\code{"character"}: A named vector of sequences.
\code{"data.frame"} or \code{"table"}: A data frame with columns for the sequence names and sequence text, as produced by \code{phylotools::read.fasta}.
\code{"DNAStringSet"}: A \code{Biostrings::DNAStringSet} object. (Default)
\code{"AAStringSet"}: A \code{Biostrings::AAStringSet} object.}

\item{verbose}{A logical argument specifying whether to display the summary output. Defaults to \code{TRUE}.}
}
\value{
Returns sequence data in the format specified by the \code{format} argument.
}
\description{
Read in a FASTA biological sequence file
}
\details{
This function is intended as a user-friendly wrapper for \code{phylotools::read.fasta},
    and converts the output into one of several formats in the local environment.
}
\examples{
write.fasta(anole.ND2, "anole.nd2.fna")

seqs <- read.fasta("anole.nd2.fna")

}
