% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.for.ncbi.ids.R
\name{search.for.ncbi.ids}
\alias{search.for.ncbi.ids}
\title{Search NCBI Entrez database for sequence IDs}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
search.for.ncbi.ids(query, database = "nuccore", verbose = TRUE, ...)
}
\arguments{
\item{query}{A character search term.}

\item{database}{A character name of an NCBI database. Defaults to \code{"nuccore"}, or the non-redundant nucleotide database (NR).
For a complete list of options run \code{rentrez::entrez_db_searchable()}}

\item{verbose}{A logical argument specifying whether to display the summary output. Defaults to \code{TRUE}.}
}
\value{
Returns an character vector of NCBI UIDs. These are the \code{esearch} object \code{ids} elements as generated by \code{rentrez::entrez_search}.
}
\description{
Search NCBI Entrez database for sequence IDs
}
\details{
This function is intended as a user-friendly wrapper for \code{rentrez::entrez_search} and accepts other arguments to that function.
}
\examples{
new.ids <- search.for.ncbi.ids("Anolis sheplani ND2")

}
