% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seventeen.species.ND2}
\alias{seventeen.species.ND2}
\title{Anole DNA sequences}
\format{
An object of class \code{DNAStringSet} of length 17.
}
\usage{
seventeen.species.ND2
}
\description{
Unaligned sequences from mitochondrial DNA of 17 species of anoles, including
    the complete CDS of NADH dehydrogenase subunit 2 (ND2), the complete tRNA-Trp sequence,
    partial sequence from tRNA-Ala.
}
\examples{
seventeen.species.ND2
names(seventeen.species.ND2)

# Make the names more minimalist
names(seventeen.species.ND2) <- sub("_ND2-tRNAtrp-tRNAala_","__",names(seventeen.species.ND2))

# Par the names down to just the species
names(seventeen.species.ND2) <- stringr::str_split_fixed(names(seventeen.species.ND2),"__",2)[,1]

# Alignment
require(msa)
seqs.msa <- msa(seventeen.species.ND2, method = "ClustalOmega", order = "aligned")
print(seqs.msa, show="complete")
}
\keyword{datasets}
