% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.alignment.R
\name{show.alignment}
\alias{show.alignment}
\title{Print a Biostrings MultipleAlignment object to the console}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
show.alignment(x, ...)
}
\arguments{
\item{x}{An object of biological sequence data. Must be one of the following:
A character string containing the FASTA format sequences as downloaded from NCBI.
A named vector of sequences.
A data frame with columns for the sequence names and sequence text, as produced by \code{phylotools::read.fasta}.
A \code{Biostrings::DNAStringSet} or \code{AAStringSet} object.}
}
\description{
Print a Biostrings MultipleAlignment object to the console
}
\details{
This function is intended as a user-friendly wrapper for \code{print(x, show="complete")} and accepts other arguments to that function.
}
\examples{
anole.alignment <- align.sequences(anole.ND2)
show.alignment(anole.alignment)

}
