% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.names.R
\name{simplify.names}
\alias{simplify.names}
\title{Get simplified names for NCBI sequences}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
simplify.names(
  x,
  separator = NULL,
  words = 2,
  abbreviate.genus = TRUE,
  prefix = "",
  postfix = ""
)
}
\arguments{
\item{x}{A character vector of complex names.}

\item{separator}{A character to use to separate the components of sequence names.}

\item{words}{An integer specifying the number of words to retain. Defaults to 2.}

\item{abbreviate.genus}{A logical argument specifying whether to treat the first word as a genus name and abbreviate it to its first character.}

\item{prefix}{A character to prefix to all names.}

\item{postfix}{A character to postfix to all names.}
}
\value{
Returns an character vector.
}
\description{
Get simplified names for NCBI sequences
}
\details{
This function takes complex sequence names and shortens them.
   Ideal for taking just the gene or species names that begin longer identifiers.
   If the separator is not specified then the function will attempt to detect one.
}
\examples{
simplify.names(anole.tree$tip.label)
simplify.names(anole.tree$tip.label, postfix = "ND2")
}
