% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.sequence.names.R
\name{simplify.sequence.names}
\alias{simplify.sequence.names}
\title{Get simplified names for NCBI sequences}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
simplify.sequence.names(
  ids,
  database = "nuccore",
  sequence.name = "",
  include.species.name = TRUE,
  include.accession = TRUE,
  separator = "_",
  ...
)
}
\arguments{
\item{ids}{A character vector of valid NCBI UIDs or accession numbers.}

\item{database}{A character name of an NCBI database. Defaults to \code{"nuccore"}, or the non-redundant nucleotide database (NR).
For a complete list of options run \code{rentrez::entrez_db_searchable()}}

\item{sequence.name}{An optional sequence name.}

\item{include.species.name}{A logical argument specifying whether to include the species name in the sequence name. Defaults to \code{TRUE}.}

\item{include.accession}{A logical argument specifying whether to include the GenBank Accession number in the sequence name. Defaults to \code{TRUE}.}

\item{separator}{A character to use to separate the components of sequence names. Defaults to the underscore.}
}
\value{
Returns an character vector of NCBI IDs. These are the \code{esearch} object \code{ids} elements as generated by \code{rentrez::entrez_search}.
}
\description{
Get simplified names for NCBI sequences
}
\details{
This function is intended as a user-friendly wrapper for \code{rentrez::entrez_summary} and accepts other arguments to that function.
}
\examples{
accession.ids <- c("EF591774.1", "AF055971.2", "AF055967.2", "AF055927.2", "AY296172.1", "AF294303.1", "AF294297.1", "AF055976.2", "AF055945.2", "AF294317.1", "AY296196.1", "AF055966.2", "AF055939.2", "AF055940.2", "AY296195.1", "AY263052.1", "AY296163.1")
short.names <- simplify.sequence.names(ids  = accession.ids, sequence.name = "ND2")
}
