% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sternites}
\alias{sternites}
\title{Sternite curvature in milkweed bugs}
\format{
An object of class \code{data.frame} with 542 rows and 5 columns.
}
\source{
Just, J., Laslo, M., Lee, Y.J., Yarnell, M., Zhang, Z. and Angelini, D.R. 2023. Distinct developmental mechanisms influence sexual dimorphisms in the milkweed bug \emph{Oncopeltus fasciatus}. \emph{Proc. R. Soc. B.} 290(1992):20222083.
    \url{https://doi.org/10.1098/rspb.2022.2083}
}
\usage{
sternites
}
\description{
Data on the size and sternite curvature of 542 milkweed bugs (\emph{Oncopeltus fasciatus}).
    Sternite curvature in this species is sex-biased, but shows extensive overlap.
    On average females have more curved sternites.
    This dataset includes measurements of individuals treated during juvenile-to-adult development with dsRNAs
    targeting several candidate genes. These genes were chosen based on potential roles in the development of
    somatic sexually dimorphism in insects.
    Curvature was calculated as the square-root of the mean of squared residuals (residual mean standard deviation)
    from the linear regression of nine landmarks aligned using generalized Procrustes analysis.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)

sternites \%>\%
  filter(!grepl("^dmrt",treatment)) \%>\%
  ggplot(aes(x = curvature, fill = sex)) +
  theme_bw() +
  facet_grid(treatment~.) +
  geom_histogram()

}
\keyword{datasets}
