% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wing.morph.crosses}
\alias{wing.morph.crosses}
\title{Crosses examining inheritance of wing morphs in the soapberry bug \emph{Jadera haematoloma}}
\format{
An object of class \code{data.frame} with 32 rows and 3 columns.
}
\source{
Fawcett MM, Parks MC, Tibbetts AE, Swart JS, Richards EM, Vanegas JC, Cenzer M, Crowley L, Simmons WR, Hou W, Angelini DR. Manipulation of insulin signaling phenocopies evolution of a host-associated polyphenism. Nature Communications 2018 9:1699.
    \url{https://www.nature.com/articles/s41467-018-04102-1}
}
\usage{
wing.morph.crosses
}
\description{
Soapberry bugs (\emph{Jadera haematoloma}) may develop to adulthood with long or short wings.
   This dataset includes the results of 32 single-pair crosses that were conducted to test the possibility that wing morphs are in fact inherited in classical Mendelian fashion.
   For each cross, the number of surviving adult F1 offspring and their frequency of the short wing (SW) morph are listed.
   Different modes of classical inheritance would predict different SW frequencies from each cross.
   Data from Supplementary Table 2 of Fawcett et al. 2018 \emph{Nature Communications}.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)

sternites \%>\%
  filter(!grepl("^dmrt",treatment)) \%>\%
  ggplot(aes(x = curvature, fill = sex)) +
  theme_bw() +
  facet_grid(treatment~.) +
  geom_histogram()

}
\keyword{datasets}
