% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.alignment.R
\name{write.alignment}
\alias{write.alignment}
\title{Write a Biostrings MultipleAlignment object to a Phylip file}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
write.alignment(
  x,
  filename = NULL,
  clean.names = TRUE,
  truncate.names = NULL,
  sequence.line.length = 60,
  ...
)
}
\arguments{
\item{x}{an \code{MsaDNAMultipleAlignment} or \code{MsaAAMultipleAlignment} object.}

\item{filename}{A character with the path and filename to create on disk. Set to \code{NULL} for display to the screen.}

\item{clean.names}{A logical argument specifying whether to make the sequence names compatible with Phylip format, replacing spaces and punctuation with underscores Defaults to \code{TRUE}.}

\item{truncate.names}{An optional number of characters by which to truncate sequence names.}

\item{sequence.line.length}{The number of sequence characters to include on a line. Set to \code{0} or \code{NULL} to remove interleaving. Default is \code{60}.}
}
\description{
Write a Biostrings MultipleAlignment object to a Phylip file
}
\examples{
anole.alignment <- align.sequences(anole.ND2)

write.alignment(anole.alignment)
write.alignment(anole.alignment, filename = "anole.alignment.phy")

}
