% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.fasta.R
\name{write.fasta}
\alias{write.fasta}
\title{Create a FASTA file from sequence data}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
write.fasta(x, filename = "sequence.fasta", verbose = TRUE)
}
\arguments{
\item{x}{An object of biological sequence data. Must be one of the following:
A character string containing the FASTA format sequences as downloaded from NCBI.
A named vector of sequences.
A data frame with columns for the sequence names and sequence text, as produced by \code{phytools::read.fasta}.
A \code{Biostrings::DNAStringSet} or \code{AAStringSet} object.}

\item{filename}{A character with the path and filename to create on disk.}

\item{verbose}{A logical argument specifying whether to display confirmation. Defaults to \code{TRUE}.}
}
\description{
Create a FASTA file from sequence data
}
\examples{
write.fasta(anole.ND2, "anole.ND2.fna")
}
